style: distributes responsibilities
split one big file in two different files
